ai/openai_client.py
import os
import openai
from dotenv import load_dotenv

load_dotenv()

openai.api_key = os.getenv("OPENAI_API_KEY")

def generate_audit_comment(issue):
    prompt = f"""
    You are an audit assistant.
    Explain this issue in professional, human-friendly language.

    Issue: {issue}
    """

    response = openai.ChatCompletion.create(
        model="gpt-4o-mini",
        messages=[{"role": "user", "content": prompt}],
        temperature=0.3
    )

    return response.choices[0].message.content
